app/requirements.txt
ruamel.yaml==0.15.35
docker==2.7.0